requirements.txt
openai>=0.28.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
